pyproject.toml
[project]
name = "uppercase"
version = "0.1.0"
description = ""
authors = [
    {name = "Alfir Tyakin",email = "alfir7777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
pytest-django = "^4.10.0"
flake8-import-order = "^0.18.2"

